Web application client made using NextJS
Added the following:

Configuration to display on the client images from remote sources.
DTO's to display the results of a search request and the information of an auction.
Tailwind CSS configuration.
Actions that make requests to the API.
Implemented application's state management using Zustand. This state manages information the different parameters used in the search request

Implemented the following components:

Listings, AppPagination components used to display all the auctions returned by a search request.
AuctionCard, CardImage, CountdownTimer components to render each auction's information.
SearchBar and Filters add functionality that allows the user to set the a search term, filter, and how to order the auctions retrieved from the API.
Logo component implements functionality that allows it to reset all the application's state.
Navbar component adds a navbar for the application where we put the logo and an input to set a search term.
EmptyFilter to show them whenever a query returns no results.