Create bid, see bids for an auction, and SignalR client that receives notifications from notification service implemented
Components to create bid form, and list bids with its information added.
Information for loaded auctions and bids retained in application's state instead of component's state.
SignalR provider connection, configuration to receives notifications from the notification server and modify the application's state with the data received from the notifications service.
Highest bid on the auction store will only be updated when the new bid is higher than the current highest.
Application's layout updated to wrap the entire web application inside the SignalR provider.

Toasts added when an auction is placed or finishes